feat(hre):add new simulator
