--- packages/code-du-travail-data/dataset/preavis-demission/package.json
@@ -1,13 +1,0 @@
-{
-  "name": "@cdt/data...preavis-demission",
-  "version": "4.1.0",
-  "private": true,
-  "main": "index.js",
-  "license": "MIT",
-  "scripts": {
-    "start": "node index.js > ./data.json"
-  },
-  "dependencies": {
-    "google-spreadsheets": "^2.0.0"
-  }
-}
