--- packages/code-du-travail-data/dataset/preavis-licenciement/CHANGELOG.md
@@ -1,167 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-# [4.1.0](https://github.com/SocialGouv/code-du-travail-numerique/compare/v4.1.0-alpha.2...v4.1.0) (2020-01-16)
-
-**Note:** Version bump only for package @cdt/data...preavis-licenciement
-
-
-
-
-
-# [4.1.0-alpha.2](https://github.com/SocialGouv/code-du-travail-numerique/compare/v4.1.0-alpha.1...v4.1.0-alpha.2) (2020-01-15)
-
-**Note:** Version bump only for package @cdt/data...preavis-licenciement
-
-
-
-
-
-# [4.1.0-alpha.1](https://github.com/SocialGouv/code-du-travail-numerique/compare/v4.1.0-alpha.0...v4.1.0-alpha.1) (2020-01-13)
-
-
-### Features
-
-* **wizzard:** allow to skip steps ([#2191](https://github.com/SocialGouv/code-du-travail-numerique/issues/2191)) ([2e574e0](https://github.com/SocialGouv/code-du-travail-numerique/commit/2e574e0e1b7b1c437253e99b7c1fd6ab3693a8cd))
-
-
-
-
-
-# [4.1.0-alpha.0](https://github.com/SocialGouv/code-du-travail-numerique/compare/v4.0.0...v4.1.0-alpha.0) (2020-01-08)
-
-**Note:** Version bump only for package @cdt/data...preavis-licenciement
-
-
-
-
-
-# [4.0.0](https://github.com/SocialGouv/code-du-travail-numerique/compare/v4.0.0-gamma.4...v4.0.0) (2020-01-07)
-
-**Note:** Version bump only for package @cdt/data...preavis-licenciement
-
-
-
-
-
-# [4.0.0-gamma.4](https://github.com/SocialGouv/code-du-travail-numerique/compare/v4.0.0-gamma.3...v4.0.0-gamma.4) (2020-01-06)
-
-**Note:** Version bump only for package @cdt/data...preavis-licenciement
-
-
-
-
-
-# [4.0.0-gamma.3](https://github.com/SocialGouv/code-du-travail-numerique/compare/v4.0.0-gamma.2...v4.0.0-gamma.3) (2020-01-03)
-
-**Note:** Version bump only for package @cdt/data...preavis-licenciement
-
-
-
-
-
-# [4.0.0-gamma.2](https://github.com/SocialGouv/code-du-travail-numerique/compare/v4.0.0-gamma.1...v4.0.0-gamma.2) (2020-01-02)
-
-**Note:** Version bump only for package @cdt/data...preavis-licenciement
-
-
-
-
-
-# [4.0.0-gamma.1](https://github.com/SocialGouv/code-du-travail-numerique/compare/v4.0.0-gamma.0...v4.0.0-gamma.1) (2019-12-31)
-
-**Note:** Version bump only for package @cdt/data...preavis-licenciement
-
-
-
-
-
-# [4.0.0-gamma.0](https://github.com/SocialGouv/code-du-travail-numerique/compare/v4.0.0-alpha.1...v4.0.0-gamma.0) (2019-12-31)
-
-**Note:** Version bump only for package @cdt/data...preavis-licenciement
-
-
-
-
-
-# 4.0.0-alpha.1 (2019-12-31)
-
-
-### Bug Fixes
-
-* **idcc:** support new kali-data  number format for idcc ([#2049](https://github.com/SocialGouv/code-du-travail-numerique/issues/2049)) ([12a8c2b](https://github.com/SocialGouv/code-du-travail-numerique/commit/12a8c2b61f8816354f8f53e26e099508459a0897)), closes [#2041](https://github.com/SocialGouv/code-du-travail-numerique/issues/2041)
-
-
-
-
-
-# 4.0.0-alpha.0 (2019-12-31)
-
-
-### Bug Fixes
-
-* **idcc:** support new kali-data  number format for idcc ([#2049](https://github.com/SocialGouv/code-du-travail-numerique/issues/2049)) ([12a8c2b](https://github.com/SocialGouv/code-du-travail-numerique/commit/12a8c2b61f8816354f8f53e26e099508459a0897)), closes [#2041](https://github.com/SocialGouv/code-du-travail-numerique/issues/2041)
-
-
-
-
-
-# [3.10.0](https://github.com/SocialGouv/code-du-travail-numerique/compare/v3.9.0...v3.10.0) (2019-12-20)
-
-**Note:** Version bump only for package @cdt/data...preavis-licenciement
-
-
-
-
-
-# 3.9.0 (2019-12-18)
-
-
-### Bug Fixes
-
-* **deps:** update dependency next to ^9.1.5 ([#1925](https://github.com/SocialGouv/code-du-travail-numerique/issues/1925)) ([c1aa17e](https://github.com/SocialGouv/code-du-travail-numerique/commit/c1aa17e12a296851194f3d709ec9d1c8e7b1e720))
-
-
-
-
-
-# 3.8.0 (2019-12-03)
-
-**Note:** Version bump only for package @cdt/data...preavis-licenciement
-
-
-
-
-
-# [3.7.0](https://github.com/SocialGouv/code-du-travail-numerique/compare/v3.6.2...v3.7.0) (2019-11-21)
-
-**Note:** Version bump only for package @cdt/data...preavis-licenciement
-
-
-
-
-
-## [3.6.2](https://github.com/SocialGouv/code-du-travail-numerique/compare/v3.6.1...v3.6.2) (2019-11-18)
-
-**Note:** Version bump only for package @cdt/data...preavis-licenciement
-
-
-
-
-
-## [3.6.1](https://github.com/SocialGouv/code-du-travail-numerique/compare/v3.6.0...v3.6.1) (2019-11-18)
-
-**Note:** Version bump only for package @cdt/data...preavis-licenciement
-
-
-
-
-
-# 3.6.0 (2019-11-18)
-
-
-### Bug Fixes
-
-* **deps:** update dependency date-fns to ^2.7.0 ([#1662](https://github.com/SocialGouv/code-du-travail-numerique/issues/1662)) ([c77414c](https://github.com/SocialGouv/code-du-travail-numerique/commit/c77414c))
