+++ packages/code-du-travail-data/dataset/simulateurs/preavis-demission.data.json
@@ -1,6 +1,10 @@
 {
   "questions": [
     {
+      "name": "name",
+      "question": "question"
+    },
+    {
       "name": "aéroport paris",
       "question": "L'entreprise appliquait elle la convention collective régionale \"Manutention et nettoyage sur les aéroports (région parisienne)\" ?"
     },
