+++ packages/code-du-travail-data/dataset/simulateurs/preavis-licenciement.data.json
@@ -1,6 +1,10 @@
 {
   "questions": [
     {
+      "name": "name",
+      "question": "question"
+    },
+    {
       "name": "aéroport paris",
       "question": "L'entreprise appliquait elle la convention collective régionale \"Manutention et nettoyage sur les aéroports (région parisienne)\" ?"
     },
@@ -1460,8 +1464,7 @@
     {
       "criteria": {
         "position": "1| I",
-        "ancienneté": "36| moins de 2 ans",
-        "âge": "1| moins de 50 ans"
+        "ancienneté": "36| moins de 2 ans"
       },
       "type": "préavis de licenciement",
       "idcc": 650,
@@ -1473,8 +1476,7 @@
     {
       "criteria": {
         "position": "1| I",
-        "ancienneté": "40| 2 ans ou plus",
-        "âge": "1| moins de 50 ans"
+        "ancienneté": "40| 2 ans ou plus"
       },
       "type": "préavis de licenciement",
       "idcc": 650,
@@ -1485,8 +1487,7 @@
     },
     {
       "criteria": {
-        "position": "3| II",
-        "âge": "1| moins de 50 ans"
+        "position": "3| II"
       },
       "type": "préavis de licenciement",
       "idcc": 650,
@@ -1497,8 +1498,7 @@
     },
     {
       "criteria": {
-        "position": "4| III",
-        "âge": "1| moins de 50 ans"
+        "position": "4| III"
       },
       "type": "préavis de licenciement",
       "idcc": 650,
@@ -1509,8 +1509,7 @@
     },
     {
       "criteria": {
-        "ancienneté": "22| moins de 1 an",
-        "âge": "2| 50 ans à 55 ans"
+        "ancienneté": "22| moins de 1 an"
       },
       "type": "préavis de licenciement",
       "idcc": 650,
@@ -1521,8 +1520,7 @@
     },
     {
       "criteria": {
-        "ancienneté": "42| entre 1 an et moins de 5 ans",
-        "âge": "2| 50 ans à 55 ans"
+        "ancienneté": "42| entre 1 an et moins de 5 ans"
       },
       "type": "préavis de licenciement",
       "idcc": 650,
@@ -1533,8 +1531,7 @@
     },
     {
       "criteria": {
-        "ancienneté": "46| 5 ans et plus",
-        "âge": "2| 50 ans à 55 ans"
+        "ancienneté": "46| 5 ans et plus"
       },
       "type": "préavis de licenciement",
       "idcc": 650,
@@ -1544,9 +1541,7 @@
       "refUrl": "https://www.legifrance.gouv.fr/affichIDCCArticle.do;jsessionid=3A4E814FD2154397637BC31D67BD3E6F.tplgfr26s_3?idArticle=KALIARTI000005836199&cidTexte=KALITEXT000005673470&dateTexte=29990101&categorieLien=id"
     },
     {
-      "criteria": {
-        "âge": "4| 55 ans et plus"
-      },
+      "criteria": {},
       "type": "préavis de licenciement",
       "idcc": 650,
       "answer": "6 mois",
@@ -6258,8 +6253,7 @@
     },
     {
       "criteria": {
-        "ancienneté": "36| moins de 2 ans",
-        "âge": "3| moins de 55 ans"
+        "ancienneté": "36| moins de 2 ans"
       },
       "type": "préavis de licenciement",
       "idcc": 2609,
@@ -6270,8 +6264,7 @@
     },
     {
       "criteria": {
-        "ancienneté": "40| 2 ans ou plus",
-        "âge": "3| moins de 55 ans"
+        "ancienneté": "40| 2 ans ou plus"
       },
       "type": "préavis de licenciement",
       "idcc": 2609,
@@ -6282,8 +6275,7 @@
     },
     {
       "criteria": {
-        "ancienneté": "48| Au moins 15 ans d'ancienneté",
-        "âge": "5| plus de 55 ans"
+        "ancienneté": "48| Au moins 15 ans d'ancienneté"
       },
       "type": "préavis de licenciement",
       "idcc": 2609,
@@ -6294,8 +6286,7 @@
     },
     {
       "criteria": {
-        "ancienneté": "47| moins de 15 ans d'ancienneté",
-        "âge": "5| plus de 55 ans"
+        "ancienneté": "47| moins de 15 ans d'ancienneté"
       },
       "type": "préavis de licenciement",
       "idcc": 2609,
@@ -6306,8 +6297,7 @@
     },
     {
       "criteria": {
-        "ancienneté": "36| moins de 2 ans",
-        "âge": "3| moins de 55 ans"
+        "ancienneté": "36| moins de 2 ans"
       },
       "type": "préavis de licenciement",
       "idcc": 2614,
@@ -6318,8 +6308,7 @@
     },
     {
       "criteria": {
-        "ancienneté": "40| 2 ans ou plus",
-        "âge": "3| moins de 55 ans"
+        "ancienneté": "40| 2 ans ou plus"
       },
       "type": "préavis de licenciement",
       "idcc": 2614,
@@ -6330,8 +6319,7 @@
     },
     {
       "criteria": {
-        "ancienneté": "48| Au moins 15 ans d'ancienneté",
-        "âge": "4| 55 ans et plus"
+        "ancienneté": "48| Au moins 15 ans d'ancienneté"
       },
       "type": "préavis de licenciement",
       "idcc": 2614,
@@ -6342,8 +6330,7 @@
     },
     {
       "criteria": {
-        "ancienneté": "47| moins de 15 ans d'ancienneté",
-        "âge": "4| 55 ans et plus"
+        "ancienneté": "47| moins de 15 ans d'ancienneté"
       },
       "type": "préavis de licenciement",
       "idcc": 2614,
