+++ packages/code-du-travail-frontend/src/outils/DureePreavisLicenciement/steps/Result.js
@@ -2,7 +2,8 @@ import React from "react";
 import PropTypes from "prop-types";
 import Link from "next/link";
 import { Toast } from "@socialgouv/react-ui";
-import data from "@cdt/data...preavis-licenciement/data.json";
+import data from "@cdt/data...simulateurs/preavis-licenciement.data.json";
+
 import { getRouteBySource, SOURCES } from "@cdt/sources";
 import { SectionTitle, Highlight } from "../../common/stepStyles";
